% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA_plot_roc.R
\name{miRNA_plotROC}
\alias{miRNA_plotROC}
\title{Plot of classifier ROC curve.}
\usage{
miRNA_plotROC(inputDataset, outputFileLabel, plotFormat = "pdf")
}
\arguments{
\item{inputDataset}{Dataset (data frame) to be used for the plot.}

\item{outputFileLabel}{Label to be used to build the name of the output file.}

\item{plotFormat}{String to set the format of the output file. Can either be 'pdf' (default) or 'png'.}
}
\value{
A ggplot object containing the plot.

This function is not exported to the package NAMESPACE, but it is called by other functions of the same package.
}
\description{
Generates a plot of the ROC curve out of the input dataset.
}
