% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA_plot_thresholds.R
\name{miRNA_plotThresholds}
\alias{miRNA_plotThresholds}
\title{Plot of scores and thresholds of a Bayes classifier.}
\usage{
miRNA_plotThresholds(
  inputDataset,
  thresholdsFrame,
  outputFileLabel,
  plotFormat = "pdf"
)
}
\arguments{
\item{inputDataset}{Dataset (data frame) to be used for the plot.}

\item{thresholdsFrame}{Diagnostic threshold values (data frame) to be used for the plot.}

\item{outputFileLabel}{Label to be used to build the name of the output file.}

\item{plotFormat}{String to set the format of the output file. Can either be 'pdf' (default) or 'png'.}
}
\value{
A ggplot object containing the plot.

This function is not exported to the package NAMESPACE, but it is called by other functions of the same package.
}
\description{
Generates a plot of the classifier scores of a dataset, as well as the corresponding classifier thresholds.
}
