% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNA_plot_histograms.R
\name{miRNA_plotHistograms}
\alias{miRNA_plotHistograms}
\title{Plot of classifier distribution histograms (Target and Versus).}
\usage{
miRNA_plotHistograms(
  inputDataset,
  thresholdFrame = character(),
  outputFileLabel,
  plotFormat = "pdf"
)
}
\arguments{
\item{inputDataset}{Dataset (data frame) to be used for the plot.}

\item{thresholdFrame}{Diagnostic threshold values (data frame) to be used for the plot. If omitted, no threshold is drawn on the histogram.}

\item{outputFileLabel}{Label to be used to build the name of the output file.}

\item{plotFormat}{String to set the format of the output file. Can either be 'pdf' (default) or 'png'.}
}
\value{
A ggplot object containing the plot.

This function is not exported to the package NAMESPACE, but it is called by other functions of the same package.
}
\description{
Generates a plot of the histograms of the score values for the target and versus sets.
}
